+++ src/loc.rs
@@ -6,7 +6,7 @@ use std::marker::PhantomData;
 
 /// An offset into the `.debug_loc` section.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct DebugLocOffset(pub u64);
+pub struct DebugLocOffset(pub usize);
 
 /// The `DebugLoc` struct represents the DWARF strings
 /// found in the `.debug_loc` section.
@@ -287,7 +287,7 @@ mod tests {
 
         let buf = section.get_contents().unwrap();
         let debug_loc = DebugLoc::<LittleEndian>::new(&buf);
-        let offset = DebugLocOffset((&first - &start) as u64);
+        let offset = DebugLocOffset((&first - &start) as usize);
         let mut locations = debug_loc.locations(offset, 4, 0x01000000).unwrap();
 
         // A normal location.
@@ -344,7 +344,7 @@ mod tests {
         assert_eq!(locations.next(), Ok(None));
 
         // An offset at the end of buf.
-        let mut locations = debug_loc.locations(DebugLocOffset(buf.len() as u64), 4, 0x01000000)
+        let mut locations = debug_loc.locations(DebugLocOffset(buf.len()), 4, 0x01000000)
             .unwrap();
         assert_eq!(locations.next(), Ok(None));
     }
@@ -378,7 +378,7 @@ mod tests {
 
         let buf = section.get_contents().unwrap();
         let debug_loc = DebugLoc::<LittleEndian>::new(&buf);
-        let offset = DebugLocOffset((&first - &start) as u64);
+        let offset = DebugLocOffset((&first - &start) as usize);
         let mut locations = debug_loc.locations(offset, 8, 0x01000000).unwrap();
 
         // A normal location.
@@ -435,7 +435,7 @@ mod tests {
         assert_eq!(locations.next(), Ok(None));
 
         // An offset at the end of buf.
-        let mut locations = debug_loc.locations(DebugLocOffset(buf.len() as u64), 8, 0x01000000)
+        let mut locations = debug_loc.locations(DebugLocOffset(buf.len()), 8, 0x01000000)
             .unwrap();
         assert_eq!(locations.next(), Ok(None));
     }
@@ -460,7 +460,7 @@ mod tests {
         assert_eq!(locations.next(), Err(Error::InvalidLocationAddressRange));
 
         // An invalid offset.
-        match debug_loc.locations(DebugLocOffset(buf.len() as u64 + 1), 4, 0x01000000) {
+        match debug_loc.locations(DebugLocOffset(buf.len() + 1), 4, 0x01000000) {
             Err(Error::UnexpectedEof) => {}
             otherwise => panic!("Unexpected result: {:?}", otherwise),
         }
